user done

+++ rental-api/src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
@@ -37,27 +37,50 @@ export class UserService {
     return this.userRepo.findOneBy({ email })
   }
 
-  async findOneById(id:number){
-    
-    return this.userRepo.findOneBy({ id })
+  /**
+   * Getting one user by id
+   * @param id - id number of number
+   * @param perms - permissions from request
+   * @returns found user
+   */
+  async findOneById(id:number, perms: string){
+    if(await this.checkPerms(id, perms))
+      return this.userRepo.findOneBy({ id })
+    throw new UnauthorizedException
   }
 
   /**
    * Updating user
    * @param id - id of user to update
+   * @param perms - permissions from request
    * @param updateUserDto - data of user to update
    * @returns result of updating user
    */
-  async update(id: number, updateUserDto: UpdateUserDto) {
-    return this.userRepo.update(id, updateUserDto)
+  async update(id: number, perms: string, updateUserDto: UpdateUserDto) {
+    if(await this.checkPerms(id, perms))
+      return this.userRepo.update(id, updateUserDto)
+    throw new UnauthorizedException
   }
 
   /**
    * Removing user
    * @param id - id of user to delete
+   * @param perms - permissions from request
    * @returns result of deleting user
    */
-  async remove(id: number) {
-    return this.userRepo.delete(id)
+  async remove(id: number, perms: string) {
+    if(await this.checkPerms(id, perms))
+      return this.userRepo.delete(id)
+    throw new UnauthorizedException
+  }
+
+  /**
+   * Permission Checker
+   * @param id - index number of user
+   * @param perms - permission from request
+   * @returns boolean if user's permissions are equal
+   */
+  async checkPerms(id: number, perms: string): Promise<boolean>{
+    return this.userRepo.findOneBy({ id }).then( user => user?.permissions == perms)
   }
 }
